get_xf_dist function, to get xf distribution from MCEq data directly
Former-commit-id: ac8090195dea0aa41f51df1428d9f2b2d0de6bb6

+++ MCEq/data_utils.py
@@ -33,7 +33,7 @@ def clean_data_directory():
             os.unlink(path.join(datad, fname))
         elif 'yields_compact' in fname:
             os.unlink(path.join(datad, fname))
-        elif fname.endswith('.ppd'):
+        elif fname.endswith('.ppd') and 'sibyll23c_aux' not in fname:
             os.unlink(path.join(datad, fname))
 
 def convert_to_compact(fname):
